requirements.txt
absl-py==2.3.1
attrs==25.3.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.1.8
click-default-group==1.2.4
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
facenet-pytorch==2.5.3
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.59.1
fsspec==2025.7.0
git-filter-repo==2.47.0
h5py==3.14.0
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
jax==0.4.30
jaxlib==0.4.30
Jinja2==3.1.6
joblib==1.5.1
keras==3.10.0
kiwisolver==1.4.7
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
mdurl==0.1.2
mediapipe==0.10.14
ml_dtypes==0.5.3
mpmath==1.3.0
mtcnn==0.1.1
mysql-connector-python==9.4.0
namex==0.1.0
networkx==3.2.1
numpy==2.0.2
opencv-contrib-python==4.12.0.88
opencv-python==4.12.0.88
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.3.1
pillow==11.3.0
pluggy==1.6.0
protobuf==4.25.8
pycparser==2.22
Pygments==2.19.2
pyparsing==3.2.3
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
rich==14.1.0
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
sounddevice==0.5.2
sqlite-fts4==1.0.3
sqlite-utils==3.38
sympy==1.13.1
tabulate==0.9.0
threadpoolctl==3.6.0
tifffile==2024.8.30
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
zipp==3.23.0
